requirements.txt
gradio>=4.0.0
python-dotenv>=1.0.0
PyPDF2>=3.0.0
nomic>=3.2.0
numpy>=1.23.0
scikit-learn>=1.2.0
python-docx>=1.0.0
openai>=1.0.0
requests>=2.28.0
nomic[local]>=2.2.0
langchain>=0.0.200